Submodule for compactness constraints
