+++ gerrychain/constraints/__init__.py
@@ -38,29 +38,15 @@ examples of this.
 
 """
 
-from .bounds import (
-    LowerBound,
-    SelfConfiguringLowerBound,
-    SelfConfiguringUpperBound,
-    UpperBound,
-    WithinPercentRangeOfBounds,
-)
-from .contiguity import (
-    contiguous,
-    contiguous_bfs,
-    no_more_discontiguous,
-    single_flip_contiguous,
-)
-from .validity import (
-    L1_polsby_popper,
-    L1_reciprocal_polsby_popper,
-    L2_polsby_popper,
-    L_minus_1_polsby_popper,
-    Validator,
-    districts_within_tolerance,
-    no_vanishing_districts,
-    no_worse_L1_reciprocal_polsby_popper,
-    no_worse_L_minus_1_polsby_popper,
-    refuse_new_splits,
-    within_percent_of_ideal_population,
-)
+from .bounds import (LowerBound, SelfConfiguringLowerBound,
+                     SelfConfiguringUpperBound, UpperBound,
+                     WithinPercentRangeOfBounds)
+from .compactness import (L1_polsby_popper, L1_reciprocal_polsby_popper,
+                          L2_polsby_popper, L_minus_1_polsby_popper,
+                          no_worse_L1_reciprocal_polsby_popper,
+                          no_worse_L_minus_1_polsby_popper)
+from .contiguity import (contiguous, contiguous_bfs, no_more_discontiguous,
+                         single_flip_contiguous)
+from .validity import (Validator, districts_within_tolerance,
+                       no_vanishing_districts, refuse_new_splits,
+                       within_percent_of_ideal_population)
