+++ gerrychain/constraints/validity.py
@@ -1,8 +1,5 @@
-import math
-
-from gerrychain.constraints.bounds import (Bounds, SelfConfiguringLowerBound,
-                                           SelfConfiguringUpperBound)
-from gerrychain.updaters import CountySplit
+from ..updaters import CountySplit
+from .bounds import Bounds
 
 
 class Validator:
@@ -45,31 +42,6 @@ class Validator:
 
         # all constraints are satisfied
         return True
-
-
-def L1_reciprocal_polsby_popper(partition):
-    return sum(1 / value for value in partition["polsby_popper"].values())
-
-
-def L1_polsby_popper(partition):
-    return sum(value for value in partition["polsby_popper"].values())
-
-
-def L2_polsby_popper(partition):
-    return math.sqrt(sum(value ** 2 for value in partition["polsby_popper"].values()))
-
-
-def L_minus_1_polsby_popper(partition):
-    return len(partition.parts) / sum(
-        1 / value for value in partition["polsby_popper"].values()
-    )
-
-
-no_worse_L_minus_1_polsby_popper = SelfConfiguringLowerBound(L_minus_1_polsby_popper)
-
-no_worse_L1_reciprocal_polsby_popper = SelfConfiguringUpperBound(
-    L1_reciprocal_polsby_popper
-)
 
 
 def within_percent_of_ideal_population(
